chore: renamed Fallback to AmenityIcon and moved it to the Amenities folder; renamed SelectedIcon to selectedIcon
