+++ FrontEnd/src/components/basics/Amenities/Amenities.tsx
@@ -6,7 +6,7 @@ import React, {
 import { amenityIcons, IconProps } from '@icons';
 import Col, { ColProps } from 'react-bootstrap/Col';
 import styles from './Amenities.module.scss';
-import FallbackIcon from '@basics/Fallback';
+import { AmenityIcon } from '@basics/Amenities';
 import cn from 'classnames';
 
 export const amenityToIcon = {
@@ -103,8 +103,8 @@ const Amenities: FunctionComponent<AmenitiesProps> = ({
   return (
     <div className={cn(styles.wrapperDefault, className)}>
       {selected.map((s) => {
-        const SelectedIcon = <FallbackIcon amenity={s} />;
-        const Icon = () => (variant !== 'onlyLabel' ? SelectedIcon : null);
+        const selectedIcon = <AmenityIcon amenity={s} />;
+        const Icon = () => (variant !== 'onlyLabel' ? selectedIcon : null);
         const Label = () => (variant !== 'onlyIcon' ? <div>{s}</div> : null);
 
         return (
