+++ FrontEnd/src/components/basics/Amenities/AmenityIcon.tsx
@@ -2,22 +2,22 @@ import React, { FunctionComponent } from 'react';
 import { amenityToIcon, Amenity } from '@basics/Amenities';
 import { Icon as IconType, miscIcons } from '@icons';
 
-interface FallbackProps {
+interface AmentiyIconProps {
   amenity: Amenity;
 }
 
 // isAmenity is a type guard to check if key is of Amenity type
-const isAmenity = (key: any): key is Amenity => {
+const isAmenityIcon = (key: any): key is Amenity => {
   return key in amenityToIcon;
 };
 
 // Fallback Icon Component returns an amenity Icon if it is in the list of amenity icons
 // otherwise, returns a fallback Icon
-const FallbackIcon: FunctionComponent<FallbackProps> = ({ amenity }) => {
-  const Icon: IconType = isAmenity(amenity)
+const AmenityIcon: FunctionComponent<AmentiyIconProps> = ({ amenity }) => {
+  const Icon: IconType = isAmenityIcon(amenity)
     ? amenityToIcon[amenity]
     : miscIcons.Fallback;
   return <Icon />;
 };
 
-export default FallbackIcon;
+export default AmenityIcon;
