+++ FrontEnd/src/components/basics/Amenities/index.ts
@@ -1,2 +1,5 @@
-export { default } from './Amenities';
+export { default as Amenities } from './Amenities';
 export * from './Amenities';
+
+export { default as AmenityIcon } from './AmenityIcon';
+export * from './AmenityIcon';
