--- FrontEnd/src/components/basics/Fallback/index.tsx
@@ -1,2 +1,0 @@
-export { default } from './Fallback';
-export * from './Fallback';
